% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccessToken.R
\name{getAccessToken}
\alias{getAccessToken}
\title{Get an access token from the Impect Customer API}
\usage{
getAccessToken(username, password)
}
\arguments{
\item{username}{your IMPECT username}

\item{password}{your IMPECT password}
}
\value{
a string containing a bearer token
}
\description{
Get an access token from the Impect Customer API
}
\examples{
\donttest{
try({ # prevent cran errors
  username <- "yourUsername"
  password <- "youPassword"
  token <- getAccessToken(username = username, password = password)
})
}
}
