% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompetitions.R
\name{getCompetitions}
\alias{getCompetitions}
\title{Return a dataframe containing all competitionIterations available to the user}
\usage{
getCompetitions(token)
}
\arguments{
\item{token}{bearer token}
}
\value{
a dataframe containing all competition iterations available to the user
}
\description{
Return a dataframe containing all competitionIterations available to the user
}
\examples{
\donttest{
try({ # prevent cran errors
  competitionIterations <- getCompetitions(token)
})
}
}
