% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventData.R
\name{getEventData}
\alias{getEventData}
\title{Return a dataframe that contains all events for a given match ID}
\usage{
getEventData(match, token)
}
\arguments{
\item{match}{Impect match ID}

\item{token}{bearer token}
}
\value{
a dataframe conatining all events for the given match ID
}
\description{
Return a dataframe that contains all events for a given match ID
}
\examples{
\donttest{
try({ # prevent cran errors
  events <- getEventData(84248, token)
})
}
}
