% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIterations.R
\name{getIterations}
\alias{getIterations}
\title{Return a dataframe containing all iterations available to the user}
\usage{
getIterations(token)
}
\arguments{
\item{token}{bearer token}
}
\value{
a dataframe containing all iterations available to the user
}
\description{
Return a dataframe containing all iterations available to the user
}
\examples{
\donttest{
try({ # prevent cran errors
  competitionIterations <- getCompetitions(token)
})
}
}
