% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchplan.R
\name{getMatchplan}
\alias{getMatchplan}
\title{Return a dataframe with basic information for all matches for a given competitionIteration}
\usage{
getMatchplan(competitionIterationId, token)
}
\arguments{
\item{competitionIterationId}{Impect competition iteration ID}

\item{token}{bearer token}
}
\value{
a dataframe containing all matches for a given competition iteration ID
}
\description{
Return a dataframe with basic information for all matches for a given competitionIteration
}
\examples{
\donttest{
try({ # prevent cran errors
  matchplan <- getMatchplan(518, token)
})
}
}
