% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchPredictions.R
\name{getMatchPredictions}
\alias{getMatchPredictions}
\title{Return a dataframe that contains match predictions for a given iteration ID}
\usage{
getMatchPredictions(iteration, token, host = "https://api.impect.com")
}
\arguments{
\item{iteration}{'IMPECT' iteration ID}

\item{token}{bearer token}

\item{host}{host environment}
}
\value{
a dataframe containing the match predictions for the given
iteration ID
}
\description{
Return a dataframe that contains match predictions for a given iteration ID
}
\examples{
# Toy example: this will error quickly (no API token)
try(match_predictions <- getMatchPredictions(
  iteration = 0,
  token = "invalid"
))

# Real usage: requires valid Bearer Token from `getAccessToken()`
\dontrun{
match_predictions <- getMatchPredictions(
  iteration = 1004,
  token = "yourToken"
)
}
}
