% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchsums.R
\name{getMatchsums}
\alias{getMatchsums}
\title{Return a dataframe that contains all player matchsums for a given match ID}
\usage{
getMatchsums(match, token)
}
\arguments{
\item{match}{Impect match ID}

\item{token}{bearer token}
}
\value{
a dataframe containing the matchsums aggregated per player and position for the given match ID
}
\description{
Return a dataframe that contains all player matchsums for a given match ID
}
\examples{
\donttest{
try({ # prevent cran errors
  matchsums <- getMatchsums(84248, token)
})
}
}
