% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerIterationAverages.R
\name{getPlayerIterationAverages}
\alias{getPlayerIterationAverages}
\title{Return a dataframe that contains all player averages for a given iteration ID}
\usage{
getPlayerIterationAverages(iteration, token)
}
\arguments{
\item{iteration}{Impect iteration ID}

\item{token}{bearer token}
}
\value{
a dataframe containing the KPI averages aggregated per player and
position for the given iteration ID
}
\description{
Return a dataframe that contains all player averages for a given iteration ID
}
\examples{
\donttest{
try({ # prevent cran errors
  matchsums <- getPlayerIterationAverages(518, token)
})
}
}
