% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerMatchSums.R
\name{getPlayerMatchsums}
\alias{getPlayerMatchsums}
\title{Return a dataframe that contains all player matchsums for a given match ID}
\usage{
getPlayerMatchsums(matches, token)
}
\arguments{
\item{matches}{Impect match IDs}

\item{token}{bearer token}
}
\value{
a dataframe containing the matchsums aggregated per player and position for the given match ID
}
\description{
Return a dataframe that contains all player matchsums for a given match ID
}
\examples{
\donttest{
try({ # prevent cran errors
  matchsums <- getPlayerMatchsums(84248, token)
})
}
}
