% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSquadIterationAverages.R
\name{getSquadIterationAverages}
\alias{getSquadIterationAverages}
\title{Return a dataframe that contains all squads averages for a given iteration ID}
\usage{
getSquadIterationAverages(iteration, token)
}
\arguments{
\item{iteration}{Impect iteration ID}

\item{token}{bearer token}
}
\value{
a dataframe containing the KPI averages aggregated per squad for the
given iteration ID
}
\description{
Return a dataframe that contains all squads averages for a given iteration ID
}
\examples{
\donttest{
try({ # prevent cran errors
  squadIterationAverages <- getSquadIterationAverages(518, token)
})
}
}
