% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSquadIterationScores.R
\name{getSquadIterationScores}
\alias{getSquadIterationScores}
\title{Return a dataframe that contains all squads scores for a given iteration ID}
\usage{
getSquadIterationScores(iteration, token)
}
\arguments{
\item{iteration}{Impect iteration ID}

\item{token}{bearer token}
}
\value{
a dataframe containing the squad scores aggregated per squad for the
given iteration ID
}
\description{
Return a dataframe that contains all squads scores for a given iteration ID
}
\examples{
\donttest{
try({ # prevent cran errors
  scores <- getSquadIterationScores(518, token)
})
}
}
