% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSquadIterationStyleOfPlay.R
\name{getSquadIterationStyleOfPlay}
\alias{getSquadIterationStyleOfPlay}
\title{Return a dataframe that contains the squad style of play values for a given
iteration ID}
\usage{
getSquadIterationStyleOfPlay(iteration, token, host = "https://api.impect.com")
}
\arguments{
\item{iteration}{'IMPECT' iteration ID}

\item{token}{bearer token}

\item{host}{host environment}
}
\value{
a dataframe containing the style of play values per squad for the
given iteration ID
}
\description{
Return a dataframe that contains the squad style of play values for a given
iteration ID
}
\examples{
# Toy example: this will error quickly (no API token)
try(squad_style_of_play <- getSquadIterationStyleOfPlay(
  iteration = 0,
  token = "invalid"
))

# Real usage: requires valid Bearer Token from `getAccessToken()`
\dontrun{
squad_style_of_play <- getSquadIterationStyleOfPlay(
  iteration = 1004,
  token = "yourToken"
)
}
}
