% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSquadMatchStyleOfPlay.R
\name{getSquadMatchStyleOfPlay}
\alias{getSquadMatchStyleOfPlay}
\title{Return a dataframe that contains the squad style of play values for the
given match IDs}
\usage{
getSquadMatchStyleOfPlay(matches, token, host = "https://api.impect.com")
}
\arguments{
\item{matches}{'IMPECT' match IDs}

\item{token}{bearer token}

\item{host}{host environment}
}
\value{
a dataframe containing the style of play values per squad for the
given match IDs
}
\description{
Return a dataframe that contains the squad style of play values for the
given match IDs
}
\examples{
# Toy example: this will error quickly (no API token)
try(squad_match_style_of_play <- getSquadMatchStyleOfPlay(
  matches = c(0, 1),
  token = "invalid"
))

# Real usage: requires valid Bearer Token from `getAccessToken()`
\dontrun{
squad_match_style_of_play <- getSquadMatchStyleOfPlay(
  matches = c(84248, 158150),
  token = "yourToken"
)
}
}
