% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSquadMatchSums.R
\name{getSquadMatchsums}
\alias{getSquadMatchsums}
\title{Return a dataframe that contains squad level matchsums for a given match ID}
\usage{
getSquadMatchsums(matches, token)
}
\arguments{
\item{matches}{Impect match IDs}

\item{token}{bearer token}
}
\value{
a dataframe containing the matchsums aggregated per squad for the given match ID
}
\description{
Return a dataframe that contains squad level matchsums for a given match ID
}
\examples{
\donttest{
try({ # prevent cran errors
  matchsums <- getSquadMatchsums(84248, token)
})
}
}
