% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSquadRatings.R
\name{getSquadRatings}
\alias{getSquadRatings}
\title{Return a dataframe that contains squad ratings for a given iteration ID}
\usage{
getSquadRatings(iteration, token)
}
\arguments{
\item{iteration}{Impect iteration ID}

\item{token}{bearer token}
}
\value{
a dataframe containing the squad ratings for the given iteration ID
}
\description{
Return a dataframe that contains squad ratings for a given iteration ID
}
\examples{
\donttest{
try({ # prevent cran errors
  ratings <- getSquadRatings(1005, token)
})
}
}
