% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubstitutions.R
\name{getSubstitutions}
\alias{getSubstitutions}
\title{Return a dataframe that contains all line-up changes for a set of given
match IDs}
\usage{
getSubstitutions(matches, token)
}
\arguments{
\item{matches}{IMPECT match ID or a list of match IDs}

\item{token}{bearer token}
}
\value{
a dataframe containing all line-up changes for a set of given
match IDs
}
\description{
Return a dataframe that contains all line-up changes for a set of given
match IDs
}
\examples{
\donttest{
try({ # prevent cran errors
  substitutions <- getSubstitutions(matches = c(84248), token = token)
})
}
}
