% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccessToken.R
\name{getAccessToken}
\alias{getAccessToken}
\title{Get an access token from the 'Impect' Customer API}
\usage{
getAccessToken(
  username,
  password,
  token_url =
    "https://login.impect.com/auth/realms/production/protocol/openid-connect/token"
)
}
\arguments{
\item{username}{your 'IMPECT' username}

\item{password}{your 'IMPECT' password}

\item{token_url}{host specific token url}
}
\value{
a string containing a bearer token
}
\description{
Get an access token from the 'Impect' Customer API
}
\examples{
# Toy example: this will error quickly (no credentials)
try(getAccessToken(username = "invalidUser", password = "invalidPassword"))

# Real usage: requires valid credentials
\dontrun{
  token <- getAccessToken(username = "yourUsername", password = "yourPassword")
}
}
