% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{Return a dataframe that contains all events and kpi values for a set of given
match IDs}
\usage{
getEvents(matches, token)
}
\arguments{
\item{matches}{IMPECT match ID or a list of match IDs}

\item{token}{bearer token}
}
\value{
a dataframe containing all events and kpi values for a set of given
match IDs
}
\description{
Return a dataframe that contains all events and kpi values for a set of given
match IDs
}
\examples{
\donttest{
try({ # prevent cran errors
  events <- getEvents(matches = c(84248), token = token)
})
}
}
