% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIterations.R
\name{getIterations}
\alias{getIterations}
\title{Return a dataframe containing all iterations available to the user}
\usage{
getIterations(token, host = "https://api.impect.com")
}
\arguments{
\item{token}{bearer token}

\item{host}{host environment}
}
\value{
a dataframe containing all iterations available to the user
}
\description{
Return a dataframe containing all iterations available to the user
}
\examples{
# Toy example: this will error quickly (no API token)
try(events <- getIterations(
  token = "invalid"
))

# Real usage: requires valid Bearer Token from `getAccessToken()`
\dontrun{
iterations <- getIterations(
  token = "yourToken"
)
}
}
