% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchplan.R
\name{getMatches}
\alias{getMatches}
\title{Return a dataframe with basic information for all matches for a given iteration ID}
\usage{
getMatches(iteration, token)
}
\arguments{
\item{iteration}{IMPECT iteration ID}

\item{token}{bearer token}
}
\value{
a dataframe containing all matches for a given iteration ID
}
\description{
Return a dataframe with basic information for all matches for a given iteration ID
}
\examples{
\donttest{
try({ # prevent cran errors
  matchplan <- getMatchplan(518, token)
})
}
}
