% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\name{getMatches}
\alias{getMatches}
\title{Return a dataframe with basic information for all matches for a given iteration ID}
\usage{
getMatches(iteration, token, host = "https://api.impect.com")
}
\arguments{
\item{iteration}{'IMPECT' iteration ID}

\item{token}{bearer token}

\item{host}{host environment}
}
\value{
a dataframe containing all matches for a given iteration ID
}
\description{
Return a dataframe with basic information for all matches for a given iteration ID
}
\examples{
# Toy example: this will error quickly (no API token)
try(matchplan <- getMatches(
  iteration = 0,
  token = "invalid"
))

# Real usage: requires valid Bearer Token from `getAccessToken()`
\dontrun{
matchplan <- getMatches(
  iteration = 1004,
  token = "yourToken"
)
}
}
