% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerProfileScores.R
\name{getPlayerProfileScores}
\alias{getPlayerProfileScores}
\title{Return a dataframe that contains all player profile scores for a given
iteration ID}
\usage{
getPlayerProfileScores(iteration, positions, token)
}
\arguments{
\item{iteration}{Impect iteration ID}

\item{positions}{list of position names}

\item{token}{bearer token}
}
\value{
a dataframe containing the player profilescores aggregated per player
for the given iteration ID and list of positions
}
\description{
Return a dataframe that contains all player profile scores for a given
iteration ID
}
\examples{
\donttest{
try({ # prevent cran errors
  scores <- getPlayerProfileScores(518, token)
})
}
}
