% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSquadRatings.R
\name{getSquadRatings}
\alias{getSquadRatings}
\title{Return a dataframe that contains squad ratings for a given iteration ID}
\usage{
getSquadRatings(iteration, token)
}
\arguments{
\item{iteration}{'IMPECT' iteration ID}

\item{token}{bearer token}
}
\value{
a dataframe containing the squad ratings for the given iteration ID
}
\description{
Return a dataframe that contains squad ratings for a given iteration ID
}
\examples{
# Toy example: this will error quickly (no API token)
try(squad_ratings <- getSquadRatings(
  iteration = 0,
  token = "invalid"
))

# Real usage: requires valid Bearer Token from `getAccessToken()`
\dontrun{
squad_ratings <- getSquadRatings(
  iteration = 1004,
  token = "yourToken"
)
}
}
