% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartingPositions.R
\name{getStartingPositions}
\alias{getStartingPositions}
\title{Return a dataframe that contains all starting positions for a set of given
match IDs}
\usage{
getStartingPositions(matches, token)
}
\arguments{
\item{matches}{IMPECT match ID or a list of match IDs}

\item{token}{bearer token}
}
\value{
a dataframe containing all starting positions for a set of given
match IDs
}
\description{
Return a dataframe that contains all starting positions for a set of given
match IDs
}
\examples{
\donttest{
try({ # prevent cran errors
  startingPositions <- getStartingPositions(matches = c(84248), token = token)
})
}
}
